+++ Content.Shared/_RMC14/Weapons/Ranged/Laser/GunToggleableLaserComponent.cs
@@ -43,13 +43,13 @@ public sealed partial class GunToggleableLaserComponent : Component
     ///     The duration multiplier to apply during aimed shot while the laser is active.
     /// </summary>
     [DataField, AutoNetworkedField]
-    public double AimDurationMultiplier = 0.6;
+    public float AimDurationMultiplier = 0.6f;
 
     /// <summary>
     ///     The value to subtract from the duration multiplier if the laser is active and the target is spotted.
     /// </summary>
     [DataField, AutoNetworkedField]
-    public double SpottedAimDurationMultiplierSubtraction = 0.15;
+    public float SpottedAimDurationMultiplierSubtraction = 0.15f;
 }
 
 [DataRecord]
